common/config: cast OPT_U32 options using uint32_t
the OPT_U32 options was translated using strict_si_cast<int>(), and then
cast the converted result to uint32_t. this could cause integer
underflow. we could have lifted the burden of checking invalid input
from the user of this option to the strict_si_cast<>() function. so in
this change, we use strict_si_cast<uint32_t>() instead, before casting
the converted value into `uint32_t`.

and a test is added accordingly

Signed-off-by: Kefu Chai <tchaikov@gmail.com>